requirements.txt
Pillow==2.6.1
click==3.3
requests==2.4.3
wsgiref==0.1.2
